only create dummy cohorts in seed file when in development mode

+++ db/seeds.rb
@@ -7,9 +7,11 @@
 #   Mayor.create(name: 'Emanuel', city: cities.first)
 # Environment variables (ENV['...']) can be set in the file .env file.
 
-Cohort.destroy_all
-Cohort.create! name: "May, 2014", start_date: "May 05, 2014"
-Cohort.create! name: "June, 2014", start_date: "June 02, 2014"
+if Rails.env.development?
+  Cohort.destroy_all
+  Cohort.create! name: "May, 2014", start_date: "May 05, 2014"
+  Cohort.create! name: "June, 2014", start_date: "June 02, 2014"
+end
 
 activities_file = ENV['FILE'] || File.join(Rails.root, "db", "activities_seed.rb")
 if File.exists?(activities_file)
